// seotracker.client/src/api/client.ts
export interface WatcherDto {
    id?: number;
    name?: string | undefined;
    searchTerm?: string | undefined;
    targetUrl?: string | undefined;
}